responders/Velociraptor/requirements.txt
cortexutils
cryptography
grpcio-tools
pyvelociraptor
thehive4py